+++ site/dat/docs/Changelog.md
@@ -4,6 +4,7 @@
  - allow using Xvfb for Selenium executor
  - don't use lynx browser when openin report links
  - add PBench executor type
+ - show scenario alias in sidebar widget for JMeter
 
 ## 0.4.5
  - load settings applied to Stepping Thread Groups